Run worktree hooks in daemon
